Avoid narrowing cast in switch case
`case -1` when the switch variable is uint32_t does not compile with C++11,
and is probably not a good idea anyway. As far as I can tell, casting -1
to unsigned gives the max unsigned value, so I replaced the -1 with UINT32_MAX.
That might be platform-dependent and I'm not even sure if it's defined
behaviour.